requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-default-group==1.2.4
cloup==0.13.1
colour==0.1.5
Cython==3.0.4
decorator==5.1.1
glcontext==2.5.0
idna==3.4
isosurfaces==0.1.0
manim==0.17.3
manim-mobject-svg==0.4.0
manim-svg-animations==0.1.2
ManimPango==0.4.4
mapbox-earcut==1.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
moderngl==5.8.2
moderngl-window==2.4.4
multipledispatch==1.0.0
networkx==2.8.8
numpy==1.26.1
Pillow==9.5.0
pycairo==1.25.0
pydub==0.25.1
pyglet==2.0.9
Pygments==2.16.1
pyobjc-core==10.0
pyobjc-framework-Cocoa==10.0
pyrr==0.10.3
requests==2.31.0
rich==13.6.0
scipy==1.11.3
screeninfo==0.8.1
skia-pathops==0.7.4
srt==3.5.3
svgelements==1.9.6
svgpathtools==1.6.1
svgwrite==1.4.3
tqdm==4.66.1
urllib3==2.0.7
watchdog==2.3.1
